docs(Dropdown|Select|Visibility): remove imports, use faker and components
