+++ docs/src/components/ComponentDoc/ComponentExample/ComponentExampleRenderSource.js
@@ -4,23 +4,13 @@ import React from 'react'
 import SourceRender from 'react-source-render'
 
 import { updateForKeys } from 'docs/src/hoc'
-import * as selectCommon from 'docs/src/examples/addons/Select/common'
-import Wireframe from 'docs/src/examples/behaviors/Visibility/Wireframe'
-import * as dropdownCommon from 'docs/src/examples/modules/Dropdown/common'
 import SearchExampleStandard from 'docs/src/examples/modules/Search/Types/SearchExampleStandard'
 import SearchExampleCategory from 'docs/src/examples/modules/Search/Types/SearchExampleCategory'
 import { babelConfig, externals } from './renderConfig'
 
-const commons = {
-  Dropdown: dropdownCommon,
-  Select: selectCommon,
-}
-
 const resolver = (importPath, { displayName }) => {
   if (externals[importPath]) return externals[importPath]
-  if (commons[displayName] && _.endsWith(importPath, '/common')) return commons[displayName]
 
-  if (_.endsWith(importPath, '/Wireframe')) return Wireframe
   if (_.endsWith(importPath, '/SearchExampleCategory')) return SearchExampleCategory
   if (_.endsWith(importPath, '/SearchExampleStandard')) return SearchExampleStandard
