--- docs/src/examples/behaviors/Visibility/Wireframe.js
@@ -1,34 +1,0 @@
-import React from 'react'
-import { Divider, Image, Segment } from 'semantic-ui-react'
-
-export default () => (
-  <Segment>
-    <Image src='/images/wireframe/centered-paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/short-paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/media-paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/centered-paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/short-paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/media-paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/short-paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/centered-paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/short-paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/media-paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/paragraph.png' />
-    <Divider />
-    <Image src='/images/wireframe/short-paragraph.png' />
-  </Segment>
-)
