+++ docs/src/examples/modules/Dropdown/Types/DropdownExampleInline.js
@@ -1,20 +1,41 @@
 import React from 'react'
 import { Dropdown } from 'semantic-ui-react'
 
-import { friendOptions } from '../common'
-// friendOptions = [
-//   {
-//     text: 'Jenny Hess',
-//     value: 'Jenny Hess',
-//     image: { avatar: true, src: '/images/avatar/small/jenny.jpg' },
-//   },
-//   ...
-// ]
+export const friendOptions = [
+  {
+    key: 'Jenny Hess',
+    text: 'Jenny Hess',
+    value: 'Jenny Hess',
+    image: { avatar: true, src: '/images/avatar/small/jenny.jpg' },
+  },
+  {
+    key: 'Elliot Fu',
+    text: 'Elliot Fu',
+    value: 'Elliot Fu',
+    image: { avatar: true, src: '/images/avatar/small/elliot.jpg' },
+  },
+  {
+    key: 'Stevie Feliciano',
+    text: 'Stevie Feliciano',
+    value: 'Stevie Feliciano',
+    image: { avatar: true, src: '/images/avatar/small/stevie.jpg' },
+  },
+  {
+    key: 'Christian',
+    text: 'Christian',
+    value: 'Christian',
+    image: { avatar: true, src: '/images/avatar/small/christian.jpg' },
+  },
+]
 
 const DropdownExampleInline = () => (
   <span>
     Show me posts by{' '}
-    <Dropdown inline options={friendOptions} defaultValue={friendOptions[0].value} />
+    <Dropdown
+      inline
+      options={friendOptions}
+      defaultValue={friendOptions[0].value}
+    />
   </span>
 )
