+++ docs/src/examples/modules/Dropdown/Types/DropdownExampleSelection.js
@@ -1,18 +1,52 @@
 import React from 'react'
 import { Dropdown } from 'semantic-ui-react'
 
-import { friendOptions } from '../common'
-// friendOptions = [
-//   {
-//     text: 'Jenny Hess',
-//     value: 'Jenny Hess',
-//     image: { avatar: true, src: '/images/avatar/small/jenny.jpg' },
-//   },
-//  ...
-// ]
+export const friendOptions = [
+  {
+    key: 'Jenny Hess',
+    text: 'Jenny Hess',
+    value: 'Jenny Hess',
+    image: { avatar: true, src: '/images/avatar/small/jenny.jpg' },
+  },
+  {
+    key: 'Elliot Fu',
+    text: 'Elliot Fu',
+    value: 'Elliot Fu',
+    image: { avatar: true, src: '/images/avatar/small/elliot.jpg' },
+  },
+  {
+    key: 'Stevie Feliciano',
+    text: 'Stevie Feliciano',
+    value: 'Stevie Feliciano',
+    image: { avatar: true, src: '/images/avatar/small/stevie.jpg' },
+  },
+  {
+    key: 'Christian',
+    text: 'Christian',
+    value: 'Christian',
+    image: { avatar: true, src: '/images/avatar/small/christian.jpg' },
+  },
+  {
+    key: 'Matt',
+    text: 'Matt',
+    value: 'Matt',
+    image: { avatar: true, src: '/images/avatar/small/matt.jpg' },
+  },
+  {
+    key: 'Justen Kitsune',
+    text: 'Justen Kitsune',
+    value: 'Justen Kitsune',
+    image: { avatar: true, src: '/images/avatar/small/justen.jpg' },
+  },
+]
 
 const DropdownExampleSelection = () => (
-  <Dropdown placeholder='Select Friend' fluid selection options={friendOptions} />
+  <Dropdown
+    placeholder='Select Friend'
+    fluid
+    selection
+    options={friendOptions}
+  />
 )
 
 export default DropdownExampleSelection
