+++ docs/src/examples/modules/Dropdown/Variations/DropdownExampleFluid.js
@@ -1,7 +1,13 @@
+import _ from 'lodash'
 import React from 'react'
 import { Dropdown, Menu } from 'semantic-ui-react'
 
-import { getOptions } from '../common'
+const getOptions = (number, prefix = 'Choice ') =>
+  _.times(number, index => ({
+    key: index,
+    text: `${prefix}${index}`,
+    value: index,
+  }))
 
 const DropdownExampleFluid = () => (
   <Menu vertical>
