+++ docs/src/examples/modules/Dropdown/Variations/DropdownExampleScrolling.js
@@ -1,7 +1,13 @@
+import _ from 'lodash'
 import React from 'react'
 import { Dropdown } from 'semantic-ui-react'
 
-import { getOptions } from '../common'
+const getOptions = (number, prefix = 'Choice ') =>
+  _.times(number, index => ({
+    key: index,
+    text: `${prefix}${index}`,
+    value: index,
+  }))
 
 const DropdownExampleScrolling = () => (
   <Dropdown placeholder='Select choice' scrolling options={getOptions(15)} />
